✨ Split modules into separate files.
